Fix: Implemented workaround for potential tensorflow bug in graph execution of multiple sequential dataset processing
When processing multiple datasets with different parameters in the preprocessors, tensorflow doesn't seem to account for parameter changes. All datasets get processed the same way the last dataset (last parameter configuration) was processed, when using class and object attributes/variables. However, implementing the changing parameters not as class/object variables but as method parameters has resolved this issue. This might be due to the graph based execution of tensorflow.